third_party/WebKit/LayoutTests/fast/canvas/canvas-getImageData-invalid-expected.txt

Test the handling of invalid arguments in canvas getImageData().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.getImageData(NaN, 10, 10, 10) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.getImageData(10, NaN, 10, 10) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.getImageData(10, 10, NaN, 10) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.getImageData(10, 10, 10, NaN) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.getImageData(Infinity, 10, 10, 10) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.getImageData(10, Infinity, 10, 10) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.getImageData(10, 10, Infinity, 10) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.getImageData(10, 10, 10, Infinity) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.getImageData(undefined, 10, 10, 10) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.getImageData(10, undefined, 10, 10) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.getImageData(10, 10, undefined, 10) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.getImageData(10, 10, 10, undefined) threw exception NotSupportedError: The implementation did not support the requested type of object or operation..
PASS ctx.getImageData(10, 10, 0, 10) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS ctx.getImageData(10, 10, 10, 0) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS successfullyParsed is true

TEST COMPLETE
